commander/requirements.txt
mcp>=1.0
aiohttp>=3.9,<4
aiogram>=3.4,<4
pyyaml>=6.0
